function_app/requirements.txt
azure-functions>=1.18.0
# PyTorch CPU-only for Azure Functions (Linux)
# Matching training env versions as closely as possible
# Using torch 2.4.1 (latest stable with good Azure compatibility)
--index-url https://download.pytorch.org/whl/cpu
torch==2.4.1
torchvision==0.19.1
Pillow==12.0.0
numpy==2.0.2
scikit-image==0.24.0